requirements.txt
coverage==4.0.3
Flask==0.12.2
Flask-RESTful==0.3.6
PyJWT==1.4.2
nose==1.3.7
pylint==1.8.4
pytest==3.5.1
pytest-cov==2.5.1
virtualenv==15.2.0
python-coveralls==2.9.1
gunicorn==19.8.1
coveralls
psycopg2
